Refactor: 인터페이스를 외부 파일에 정의
